Revert "fix taranis mappings"
This reverts commit d3799aed8c760a9a5f34cf0418f3d9b25b0dbd7a.

The problem with this change is that the Taranis now looks
to the sim than it does on hardware. With this change a Taranis
in TAER mode will look like AETR mode to the sim, but will still
look like TAER mode to the hardware. So the mapping between sim
hardware is broken. The correct solution is to configure the
Taranis to be in AETR mode.